add new module, organize packeges
