+++ .idea/compiler.xml
@@ -1,6 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="CompilerConfiguration">
-    <bytecodeTargetLevel target="17" />
+    <bytecodeTargetLevel target="11">
+      <module name="desafio-android.app" target="17" />
+    </bytecodeTargetLevel>
   </component>
 </project>
