+++ app/src/main/java/com/picpay/desafio/android/network/PicPayService.kt
@@ -1,5 +1,6 @@
-package com.picpay.desafio.android
+package com.picpay.desafio.android.network
 
+import com.picpay.desafio.android.domain.model.User
 import retrofit2.Call
 import retrofit2.http.GET
