+++ app/src/main/java/com/picpay/desafio/android/presentation/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.picpay.desafio.android
+package com.picpay.desafio.android.presentation
 
 import android.view.View
 import android.widget.ProgressBar
@@ -8,6 +8,9 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.google.gson.Gson
 import com.google.gson.GsonBuilder
+import com.picpay.desafio.android.R
+import com.picpay.desafio.android.domain.model.User
+import com.picpay.desafio.android.network.PicPayService
 import okhttp3.OkHttpClient
 import retrofit2.Call
 import retrofit2.Callback
