+++ app/src/main/java/com/picpay/desafio/android/presentation/UserListAdapter.kt
@@ -1,10 +1,11 @@
-package com.picpay.desafio.android
+package com.picpay.desafio.android.presentation
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.RecyclerView
 import com.picpay.desafio.android.databinding.ListItemUserBinding
+import com.picpay.desafio.android.domain.model.User
 
 class UserListAdapter : RecyclerView.Adapter<UserListItemViewHolder>() {
