+++ app/src/main/java/com/picpay/desafio/android/presentation/UserListDiffCallback.kt
@@ -1,6 +1,7 @@
-package com.picpay.desafio.android
+package com.picpay.desafio.android.presentation
 
 import androidx.recyclerview.widget.DiffUtil
+import com.picpay.desafio.android.domain.model.User
 
 class UserListDiffCallback(
     private val oldList: List<User>,
