+++ app/src/test/java/com/picpay/desafio/android/ExampleServiceTest.kt
@@ -2,6 +2,8 @@ package com.picpay.desafio.android
 
 import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.whenever
+import com.picpay.desafio.android.domain.model.User
+import com.picpay.desafio.android.network.PicPayService
 import junit.framework.Assert.assertEquals
 import org.junit.Test
 import retrofit2.Call
