+++ build.gradle
@@ -3,4 +3,5 @@ plugins {
     alias(libs.plugins.kotlin.android) apply false
     alias(libs.plugins.android.library) apply false
     alias(libs.plugins.ksp) apply false
+    alias(libs.plugins.jetbrains.kotlin.jvm) apply false
 }
