+++ settings.gradle
@@ -21,3 +21,4 @@ dependencyResolutionManagement {
 
 rootProject.name = "desafio-android"
 include(":app")
+include ':core'
